uri restriction filters

+++ app/Controllers/Admindashboard.php
@@ -38,11 +38,10 @@ class Admindashboard extends BaseController
 
     public function index()
     {
-
         if (!session()->has('admin_logged')) {
             return redirect()->to(base_url().'admin-login/');
         }
-
+       
         $data['requests'] = $this->requestModel->orderBy('id', 'DESC')->findAll();
         $data['total'] = $this->createEmployee->getTotal();
         $data['totalDept'] = $this->departmentModel->getTotal();
@@ -66,6 +65,9 @@ class Admindashboard extends BaseController
 
     public function employees()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
        
         $data['employees'] = $this->createEmployee->paginate(10);
         $data['pager'] = $this->createEmployee->pager;
@@ -76,6 +78,10 @@ class Admindashboard extends BaseController
 
     public function departments()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data['departments'] = $this->departmentModel->findAll();
 
         return view('department_view', $data);
@@ -125,6 +131,10 @@ class Admindashboard extends BaseController
 
     public function deleteDept($id = null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         if ($this->departmentModel->where('id', $id)->delete()) {
             session()->setTempdata('delete_dept', "Department deleted successfully");
             return redirect()->to(base_url().'admindashboard/departments');
@@ -133,6 +143,10 @@ class Admindashboard extends BaseController
 
     public function leave()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data['leaves'] = $this->leaveModel->findAll();
         return view('leave_view', $data);
     }
@@ -140,6 +154,10 @@ class Admindashboard extends BaseController
     public function newEmployee()
     {
        
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         //define rules to validate form
        $data = [];
 
@@ -216,6 +234,10 @@ class Admindashboard extends BaseController
 
     public function newDept()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         //define rules to validate form
         $data = [];
 
@@ -258,6 +280,10 @@ class Admindashboard extends BaseController
     public function newLeave()
     {
        
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         //define rules to validate form
         // jg
         $data = [];
@@ -300,6 +326,9 @@ class Admindashboard extends BaseController
    
     public function editLeave($id=null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
        
         //define rules to validate form
         $data = [];
@@ -343,6 +372,10 @@ class Admindashboard extends BaseController
 
     public function deleteLeave($id=null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         //delete leave type based on id and unset id
         if ($this->leaveModel->where('id', $id)->delete()) {
             session()->setTempdata('leave_delete', "Leave type deleted successfully.");
@@ -355,12 +388,20 @@ class Admindashboard extends BaseController
 
     public function pending()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data['requests'] = $this->requestModel->orderBy('id', 'DESC')->findAll();
         return view('pending_view', $data);
     }
 
     public function approved()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data['approved'] = $this->approvedModel->paginate(10);
         $data['pager'] = $this->approvedModel->pager;
 
@@ -368,6 +409,10 @@ class Admindashboard extends BaseController
     }
     public function declined()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data['declined'] = $this->declinedModel->paginate(10);
         $data['pager'] = $this->declinedModel->pager;
 
@@ -377,11 +422,19 @@ class Admindashboard extends BaseController
 
     public function leaveHistory()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         return view('leave_history_view');
     }
 
     public function leaveDetails($id = null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         helper('form');
 
         $data['leave'] = $this->requestModel->find($id);
@@ -478,6 +531,10 @@ class Admindashboard extends BaseController
 
     public function editEmployee($id=null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         //define rules to validate form
        $data = [];
        $data['editemployee'] = $this->createEmployee->find($id);
@@ -527,6 +584,10 @@ class Admindashboard extends BaseController
 
     public function deleteEmployee($id = null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         if ($this->createEmployee->where('id', $id)->delete()) {
             session()->setTempdata('delete_success', 'Employee deleted successfully!');
         }
@@ -536,6 +597,10 @@ class Admindashboard extends BaseController
 
     public function Admin()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data['admins'] = $this->createAdmin->paginate(10);
         $data['pager'] = $this->createAdmin->pager;
 
@@ -544,6 +609,10 @@ class Admindashboard extends BaseController
 
     public function newAdmin()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         $data = [];
 
         $rules = [
@@ -586,6 +655,10 @@ class Admindashboard extends BaseController
 
     public function logout()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         //destroy login session
         session()->remove('admin_logged');
         session()->destroy();
@@ -595,6 +668,9 @@ class Admindashboard extends BaseController
 
     public function editAdmin($id=null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
         //define rules to validate form
        $data = [];
        $data['editadmin'] = $this->createAdmin->find($id);
@@ -642,6 +718,10 @@ class Admindashboard extends BaseController
 
     public function deleteAdmin($id = null)
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
         if ($this->createAdmin->where('id', $id)->delete()) {
             session()->setTempdata('delete_success', 'HOD deleted successfully!');
         }
@@ -651,6 +731,10 @@ class Admindashboard extends BaseController
 
     public function changePassword()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+
             $data = [];
             $data['userdata'] = $this->createAdmin->getLoggedUserData(session()->get('admin_logged'));
         
@@ -702,6 +786,10 @@ class Admindashboard extends BaseController
 
     public function updateAdmin()
     {
+        if (!session()->has('admin_logged')) {
+            return redirect()->to(base_url().'admin-login/');
+        }
+        
         $data = [];
 
         $uniid = session()->get('admin_logged');
